+++ cli/cli/Commands/CheckCommands/CheckCommand.cs
@@ -1,0 +1,10 @@
+using JetBrains.Annotations;
+
+namespace cli.CheckCommands;
+
+public class CheckCommandCommandGroup : CommandGroup
+{
+    public CheckCommandCommandGroup() : base("checks", "Check if your projects for known issues ")
+    {
+    }
+}